refactor: consolidate CLI logic and simplify project structure
- Merge src/action.ts and root index.ts into src/cli.ts.
- Move entry point to src/index.ts and update package.json module field.
- Refactor Action class logic into BaseCommand within src/cli.ts for cleaner inheritance.
- Switch git operations to use Bun's $ shell for better readability and performance.
- Implement --dry-run support for all commands to safely simulate file system changes and Paratranz uploads.
