+++ src/index.ts
@@ -1,0 +1,4 @@
+import { run } from '~/cli.ts'
+
+// eslint-disable-next-line antfu/no-top-level-await
+await run()
